librbd/AioCompletion: ensure read result assembly happens last
When reading from the cache, we can have a situation where each AioRead
completes as it is added to the completion, causing the final buffer
assembly to happen before the other requests are added.  It needs to run
only after we are finished adding requests.  Move it into a helper and
ensure it is only called when pending_count == 0 *and* building == false.

Signed-off-by: Sage Weil <sage@inktank.com>