osdc/ObjectCacher: trigger copy-on-write before read
If we are trying to read from an object, and it doesn't exist, but we have
dirty buffers, and a write could trigger copy-on-write, then flush our
write first to trigger the copy-on-write before we proceed with the read.
This ensures that a pattern like

CLONE
WRITE 0~10  (into cache)
READ  10~20

will flush the write and trigger a cow so that the read will see parent
data.

Signed-off-by: Sage Weil <sage@inktank.com>