Add divergent_priors test
Tests scenario where merge_old_entry encounters a divergent
entry where the prior_version is prior to log_tail.  This
is a problem since it will go into the missing set, but won't
be re-added to the missing set during read_log() if the node
restarts prior to recovering the object.

Signed-off-by: Samuel Just <sam.just@inktank.com>