mon: process failures when osds go down
If we see an osd go down, process any pending failure_info reports we have.
Reply, and then remove the record from the map.

This ensures that we process failures and clean up regardless of *why* or
*who* did the marking down; either way, the osd was failed.

This fixes crashes like #3477, where the failure record was removed when we
updated the pending_map, but additional failures came in, and we didn't
clean up.

Signed-off-by: Sage Weil <sage@inktank.com>