rbd.py: update scratch and test image sizes
Test 167 was failing due to running out of space on the scratch
file system.  The test reserves 21MB in a file, and repeats 50
times.  It required just over 1GB, so I bumped the default size
for the testing device to 1200 MB.  I increased the test device
size as well.

This resolves http://tracker.newdream.net/issues/3864.

Signed-off-by: Alex Elder <elder@inktank.com>